pyproject.toml
[tool.poetry]
name = "inventorist"
version = "0.1.0"
description = ""
authors = ["Alex Rozum <rozumalex@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytelegrambotapi = "^3.7.3"
requests = "^2.24.0"
sqlalchemy = "^1.3.19"
pyyaml = "^5.3.1"
dotmap = "^1.3.17"
flake8 = "^3.8.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
